feat: Implement batch processing for customer and user entities
- Added ResponseBuilder for standardized API responses.
- Created CountriesController to fetch top countries based on user data.
- Developed CustomerBatchController to handle customer batch uploads.
- Introduced UserController for user-related operations, including JSON file uploads.
- Defined ResponseDTO and SuperUserResponseDTO for response structures.
- Implemented TopCountryDTO for top country data representation.
- Removed obsolete UserBatchUploadController.
- Configured CustomerJobConfig and UserJobConfig for batch processing jobs.
- Added JobCompletionNotificationListener and UserSkipListener for job monitoring and error handling.
- Created Customer and User entity classes with JPA annotations.
- Developed UserService for user-related business logic.
- Established repositories for Customer and User entities.
- Updated processors for Customer and User to handle data transformations.
