--- src/main/java/com/chanllenge/batch/entity/customer/Customer.java
@@ -1,146 +1,0 @@
-package com.chanllenge.batch.entity.customer;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Entity
-@Table(name = "customer")
-@AllArgsConstructor
-@NoArgsConstructor
-public class Customer {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-	private int index;
-    private String customerId;
-    private String firstName;
-    private String lastName;
-    private String company;
-    private String city;
-    private String country;
-    private String phone1;
-    private String phone2;
-    private String email;
-    private String subscriptionDate;
-    private String website;
-    private String fullName;
-
-    public Long getId() {
-		return id;
-	}
-    
-	public void setId(Long id) {
-		this.id = id;
-	}
-	
-	public int getIndex() {
-		return index;
-	}
-	
-	public void setIndex(int index) {
-		this.index = index;
-	}
-	
-	public String getCustomerId() {
-		return customerId;
-	}
-	
-	public void setCustomerId(String customerId) {
-		this.customerId = customerId;
-	}
-	
-	public String getFirstName() {
-		return firstName;
-	}
-	
-	public void setFirstName(String firstName) {
-		this.firstName = firstName;
-	}
-	
-	public String getLastName() {
-		return lastName;
-	}
-	
-	public void setLastName(String lastName) {
-		this.lastName = lastName;
-	}
-	
-	public String getCompany() {
-		return company;
-	}
-	
-	public void setCompany(String company) {
-		this.company = company;
-	}
-	
-	public String getCity() {
-		return city;
-	}
-	
-	public void setCity(String city) {
-		this.city = city;
-	}
-	
-	public String getCountry() {
-		return country;
-	}
-	
-	public void setCountry(String country) {
-		this.country = country;
-	}
-	
-	public String getPhone1() {
-		return phone1;
-	}
-	
-	public void setPhone1(String phone1) {
-		this.phone1 = phone1;
-	}
-	
-	public String getPhone2() {
-		return phone2;
-	}
-	
-	public void setPhone2(String phone2) {
-		this.phone2 = phone2;
-	}
-	
-	public String getEmail() {
-		return email;
-	}
-	
-	public void setEmail(String email) {
-		this.email = email;
-	}
-	
-	public String getSubscriptionDate() {
-		return subscriptionDate;
-	}
-	
-	public void setSubscriptionDate(String subscriptionDate) {
-		this.subscriptionDate = subscriptionDate;
-	}
-	
-	public String getWebsite() {
-		return website;
-	}
-	
-	public void setWebsite(String website) {
-		this.website = website;
-	}
-	
-	public String getFullName() {
-		return fullName;
-	}
-	
-	public void setFullName(String fullName) {
-		this.fullName = fullName;
-	}
-}
