--- src/main/java/com/chanllenge/batch/repository/customer/CustomerRepository.java
@@ -1,7 +1,0 @@
-package com.chanllenge.batch.repository.customer;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.chanllenge.batch.entity.customer.Customer;
-
-public interface CustomerRepository extends JpaRepository<Customer, Long> {}
